requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
audioread=3.0.1=pypi_0
blosc2=2.0.0=pypi_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.7.4=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cython=3.0.10=pypi_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.0=pypi_0
filelock=3.13.1=pypi_0
fsspec=2024.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.23.4=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.0.0=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.10.2.post1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.41.1=pypi_0
markupsafe=2.1.5=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.0=pypi_0
numba=0.58.1=pypi_0
numexpr=2.8.6=pypi_0
numpy=1.24.1=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.20.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.2.0=pypi_0
pip=24.0=py38h06a4308_0
platformdirs=4.2.2=pypi_0
pooch=1.8.2=pypi_0
protobuf=5.27.2=pypi_0
psutil=6.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.22=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-tcn=1.1.0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=2.8.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py38h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=pypi_0
tables=3.8.0=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.7=pypi_0
tk=8.6.14=h39e8969_0
torch=2.3.1+cu118=pypi_0
torchaudio=2.3.1+cu118=pypi_0
torchlibrosa=0.1.0=pypi_0
torchvision=0.18.1+cu118=pypi_0
tqdm=4.66.4=pypi_0
triton=2.3.1=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=pypi_0
wandb=0.17.4=pypi_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
